Rest server is now optional and available through a separate homeport application, the hpd_example_easy have been updated to illustrate how to start rest server now.
