+++ OperationHandler/homeport/CMakeLists.txt
@@ -37,6 +37,5 @@ include_directories(${CMAKE_CURRENT_BINARY_DIR}/include)
 include_directories(${CMAKE_CURRENT_SOURCE_DIR}/src)
 
 add_subdirectory(include)
-add_subdirectory(examples)
 add_subdirectory(src)
