+++ REST/hpdREST/include/hpd_rest.h
@@ -23,16 +23,17 @@
   The views and conclusions contained in the software and documentation are those of the
   authors and should not be interpreted as representing official policies, either expressed*/
 
-#ifndef LR_INTERFACE_H
-#define LR_INTERFACE_H
+#ifndef HOMEPORT_HPD_REST_H
+#define HOMEPORT_HPD_REST_H
 
 #include "homeport.h"
-#include <stddef.h>
 
-struct lr_request;
+struct hpd_lr {
+    struct lr *lr;
+    Listener *dev_listener;
+};
 
-int lri_registerService(HomePort *homeport, Service *service);
-int lri_unregisterService( HomePort *homeport, char* uri );
-int lri_getConfiguration(void *srv_data, void **req_data, struct lr_request *req, const char *body, size_t len);
+int hpd_rest_init(struct hpd_lr *data, HomePort *hp, int port);
+int hpd_rest_deinit(struct hpd_lr *data, HomePort *hp);
 
-#endif
+#endif //HOMEPORT_HPD_REST_H
