+++ REST/hpdREST/src/xml.c
