+++ REST/hpdREST/src/xml.h
